feat: 연동된 스토리의 상태에 따라 에픽 막대의 색상이 바뀜 [#29]
- util 함수에 필요한 함수들 이동
- EpicRenderInfo 타입 수정
