+++ client/src/components/RoadmapBars/index.tsx
@@ -5,12 +5,10 @@ import RoadmapItem from '@/components/RoadmapItem';
 import { useEpicDispatch, useEpicState, useStoryState } from '@/lib/hooks/useContextHooks';
 import { addDate } from '@/lib/utils/date';
 import { getEpicById, updateEpicById } from '@/lib/api/epic';
-import { filterStoriesAboutEpic, makeEpicRenderInfo } from '@/lib/utils/epic';
+import { makeEpicRenderInfo } from '@/lib/utils/epic';
 import { useSocketReceive, useSocketSend } from '@/lib/hooks';
 
 const COLUMNS = 15;
-
-export type RoadmapBarsStatus = 'NOT_STARTED' | 'STARTED' | 'ALL_DONE';
 
 interface RoadmapBarsProps {
   rangeFrom: Date;
@@ -23,7 +21,11 @@ const RoadmapBars = ({ rangeFrom, rangeTo, dayRow, isToday }: RoadmapBarsProps) 
   const epics = useEpicState();
   const stories = useStoryState();
   const [currentDrag, setCurrentDrag] = useState({ targetId: -1, isDraggingLeft: false });
-  const epicRenderInfo = makeEpicRenderInfo(epics, { rangeFrom, rangeTo, columns: COLUMNS });
+  const epicRenderInfo = makeEpicRenderInfo(epics, stories, {
+    rangeFrom,
+    rangeTo,
+    columns: COLUMNS,
+  });
 
   const dispatchEpic = useEpicDispatch();
   const emitUpdateEpicBar = useSocketSend('UPDATE_EPIC_BAR');
@@ -81,25 +83,14 @@ const RoadmapBars = ({ rangeFrom, rangeTo, dayRow, isToday }: RoadmapBarsProps) 
     emitUpdateEpicBar(currentDrag.targetId);
   };
 
-  const getStatusFromStories = ({
-    todos,
-    inProgresses,
-    dones,
-  }: ReturnType<typeof filterStoriesAboutEpic>): RoadmapBarsStatus => {
-    if (todos.length === 0 && inProgresses.length === 0) return 'ALL_DONE';
-    else if (inProgresses.length === 0 && dones.length === 0) return 'NOT_STARTED';
-    else return 'STARTED';
-  };
-
   return (
     <>
       <S.Container>
-        {epicRenderInfo.map(({ id, length, exceedsLeft, exceedsRight, index }) => {
-          const filteredStories = filterStoriesAboutEpic(id, stories);
+        {epicRenderInfo.map(({ id, length, exceedsLeft, exceedsRight, index, status }) => {
           return (
             <RoadmapItem
               key={id}
-              status={getStatusFromStories(filteredStories)}
+              status={status}
               columns={COLUMNS}
               index={index}
               length={length}
